+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add a workaround for overlapping events when using ns resolution traces (https://github.com/pytorch/pytorch/pull/122425)
 - Better handling of CUDA sync events with steam = -1
 - (Experimental) Added save and restore feature for critical path graph.
+- Add nccl collective fields to parser config
 
 #### Changed
 - Change test data path in unittests from relative path to real path to support running test within IDEs.